% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nswarupLib.R
\name{linear.regression}
\alias{linear.regression}
\title{Linear Regression}
\usage{
linear.regression(xVector, yVector, numCols)
}
\arguments{
\item{xVector}{Vector with independent variable}

\item{yVector}{Vector with dependent variable}

\item{numCols}{The number of colums you want. Standard is one column of 1's and another that is  xVector}
}
\value{
A vector of betas, where the first cell is your y-int and the second cell represents your beta - "the slope"
}
\description{
Finds betas
}
