% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nswarupLib.R
\name{monthly.payments}
\alias{monthly.payments}
\title{Calculate the monthly payments of a loan}
\usage{
monthly.payments(p, rm, n)
}
\arguments{
\item{p}{initial loan balance}

\item{rm}{monthly rate}

\item{n}{the number of payments expected}
}
\value{
The static monthly payment for a loan
}
\description{
Calculate the monthly payments of a loan
}
