% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nswarupLib.R
\name{new.mat}
\alias{new.mat}
\title{To build a matrix with our info}
\usage{
new.mat(n, p, mp, rm)
}
\arguments{
\item{n}{the number of payments expected}

\item{p}{initial loan balance}

\item{mp}{monthly payments is  xVector}
}
\value{
Returns a matrix with the info I want
}
\description{
To build a matrix with our info
}
